extract workbench rebalance snapshot parser
